requirements.txt
requests==2.31.0
pyserial==3.5